LayoutTests/platform/ios-simulator/fast/dom/Geolocation/page-reload-cancel-permission-requests-expected.txt

Tests that when a page is reloaded, the frame is properly detached from the Geolocation object to ensure that no permission requests are in progress.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL numPendingRequests should be 0. Was 1.
PASS successfullyParsed is true

TEST COMPLETE
